+++ R/load_params.R
@@ -36,8 +36,9 @@ load_params = function(topic = c("sleep", "metrics", "rawdata",
                         nap_markerbutton_max_distance = 30,
                         SRI1_smoothing_wsize_hrs = NULL,
                         SRI1_smoothing_frac = NULL,
-                        spt_min_block =  30,
-                        spt_max_gap =  60,
+                        spt_min_block_dur =  30,
+                        spt_max_gap_dur =  60,
+                        spt_max_gap_ratio = NULL,
                         HorAngle_threshold = 60)
   }
   if ("metrics" %in% topic) {