+++ NEWS.md
@@ -2,9 +2,9 @@
 
 - Part 3:
 
-  - Add parameters "spt_min_block", "spt_max_gap", and "HorAngle_threshold" to improve control over algorithms used in part 3 #1295
+  - Add parameters "spt_min_block_dur", "spt_max_gap_dur", "spt_max_gap_ratio", and "HorAngle_threshold" to improve control over algorithms used in part 3 #1295
 
-  - Re-factor code corresponding to final 3 steps in HDCZA and HorAngle algorithms. As part of this effort a 1 epoch incorrect offset in window end was fixed #1295
+  - Re-factor code corresponding to final 3 steps in HDCZA and HorAngle algorithms. As part of this effort a 1 epoch incorrect offset in window-end was fixed #1295
 
 # CHANGES IN GGIR VERSION 3.2-6